+++ udpduplex/src/udpduplex.cpp
@@ -154,14 +154,14 @@ void UDPDuplex::setClientHostName(const std::string &hostName)
     }
 }
 
-void UDPDuplex::setClientTimeout(unsigned int timeout)
+void UDPDuplex::setClientTimeout(unsigned long int timeout)
 {
     if ((this->m_udpObjectType  == UDPObjectType::UDP_CLIENT) || (this->m_udpObjectType  == UDPObjectType::UDP_DUPLEX)) {
         this->m_udpClient->setTimeout(timeout);   
     }
 }
 
-void UDPDuplex::setTimeout(unsigned int timeout)
+void UDPDuplex::setTimeout(unsigned long int timeout)
 {
     if (this->m_udpObjectType  == UDPObjectType::UDP_SERVER) {
         this->m_udpServer->setTimeout(timeout);
@@ -191,7 +191,7 @@ std::string UDPDuplex::clientHostName() const
     }
 }
 
-unsigned int UDPDuplex::clientTimeout() const
+unsigned long int UDPDuplex::clientTimeout() const
 {
     if ((this->m_udpObjectType  == UDPObjectType::UDP_CLIENT) || (this->m_udpObjectType  == UDPObjectType::UDP_DUPLEX)) {
         return this->m_udpClient->timeout();
@@ -209,7 +209,7 @@ uint16_t UDPDuplex::serverPortNumber() const
     }
 }
 
-unsigned int UDPDuplex::serverTimeout() const
+unsigned long int UDPDuplex::serverTimeout() const
 {
     if ((this->m_udpObjectType  == UDPObjectType::UDP_SERVER) || (this->m_udpObjectType  == UDPObjectType::UDP_DUPLEX)) {
         return this->m_udpServer->timeout();
@@ -225,14 +225,14 @@ void UDPDuplex::setServerPortNumber(uint16_t portNumber)
     }
 }
 
-void UDPDuplex::setServerTimeout(unsigned int timeout)
+void UDPDuplex::setServerTimeout(unsigned long int timeout)
 {
     if ((this->m_udpObjectType  == UDPObjectType::UDP_SERVER) || (this->m_udpObjectType  == UDPObjectType::UDP_DUPLEX)) {
         this->m_udpServer->setTimeout(timeout);
     }
 }
 
-unsigned int UDPDuplex::timeout() const
+unsigned long int UDPDuplex::timeout() const
 {
     if (this->m_udpObjectType  == UDPObjectType::UDP_SERVER) {
         return this->m_udpServer->timeout();
